pkg: Upgrade jest to 27
