+++ package.json
@@ -30,7 +30,7 @@
     "@anansi/babel-preset": "2.7.4",
     "@anansi/browserslist-config": "1.3.0",
     "@anansi/eslint-plugin": "0.11.15",
-    "@anansi/jest-preset": "^0.2.2",
+    "@anansi/jest-preset": "^0.4.12",
     "@babel/cli": "7.14.5",
     "@babel/core": "7.14.6",
     "@babel/node": "^7.14.2",
@@ -40,11 +40,10 @@
     "@testing-library/react": "^12.0.0",
     "@testing-library/react-hooks": "~7.0.0",
     "@testing-library/react-native": "^7.2.0",
-    "@types/jest": "^26.0.20",
+    "@types/jest": "^26.0.23",
     "@types/lodash": "^4.14.166",
     "@types/react": "^17.0.0",
     "@zerollup/ts-transform-paths": "^1.7.18",
-    "babel-jest": "^26.6.3",
     "benchmark": "^2.1.4",
     "conventional-changelog-anansi": "^0.2.0",
     "copyfiles": "^2.4.1",
@@ -60,7 +59,7 @@
     "eslint-plugin-react": "^7.22.0",
     "eslint-plugin-react-hooks": "^4.2.0",
     "ignore-styles": "^5.0.1",
-    "jest": "^26.6.3",
+    "jest": "^27.0.5",
     "lerna": "^4.0.0",
     "mkdirp": "^1.0.4",
     "nock": "^13.0.6",
