+++ packages/core/src/react-integration/__tests__/integration-endpoint.web.tsx
@@ -575,319 +575,5 @@ for (const makeProvider of [makeCacheProvider, makeExternalCacheProvider]) {
         ),
       ).toEqual([5, 3, 7, 8]);
     });
-
-    describe('Optimistic Updates', () => {
-      it('works with partial update', async () => {
-        const params = { id: payload.id };
-        mynock.patch('/article-cooler/5').reply(200, {
-          ...payload,
-          title: 'some other title',
-          content: 'real response',
-        });
-
-        const { result, waitForNextUpdate } = renderRestHook(
-          () => {
-            const put = useFetcher(CoolerArticleResource.partialUpdate());
-            const article = useCache(CoolerArticleResource.detail(), params);
-            // @ts-expect-error
-            article.doesnotexist;
-            return { put, article };
-          },
-          {
-            results: [
-              {
-                request: CoolerArticleResource.detail(),
-                params,
-                result: payload,
-              },
-            ],
-          },
-        );
-        expect(result.current.article).toEqual(
-          CoolerArticleResource.fromJS(payload),
-        );
-        const promise = result.current.put(params, { content: 'changed' });
-        expect(result.current.article).toBeInstanceOf(CoolerArticleResource);
-        expect(result.current.article).toEqual(
-          CoolerArticleResource.fromJS({
-            ...payload,
-            content: 'changed',
-          }),
-        );
-        await promise;
-        expect(result.current.article).toEqual(
-          CoolerArticleResource.fromJS({
-            ...payload,
-            title: 'some other title',
-            content: 'real response',
-          }),
-        );
-      });
-
-      it('works with deletes', async () => {
-        const params = { id: payload.id };
-        mynock.delete('/article-cooler/5').reply(200, '');
-
-        const { result, waitForNextUpdate } = renderRestHook(
-          () => {
-            const del = useFetcher(CoolerArticleResource.delete());
-            const articles = useCache(CoolerArticleResource.list(), {});
-            return { del, articles };
-          },
-          {
-            results: [
-              {
-                request: CoolerArticleResource.list(),
-                params: {},
-                result: [payload],
-              },
-            ],
-          },
-        );
-        expect(result.current.articles).toEqual([
-          CoolerArticleResource.fromJS(payload),
-        ]);
-        const promise = result.current.del(params);
-        expect(result.current.articles).toEqual([]);
-        await promise;
-        expect(result.current.articles).toEqual([]);
-      });
-
-      it('works with eager creates', async () => {
-        const body = { id: -1111111111, content: 'hi' };
-        const existingItem = ArticleResourceWithOtherListUrl.fromJS({
-          id: 100,
-          content: 'something',
-        });
-
-        mynock.post(`/article/`).reply(201, {
-          ...payload,
-          title: 'some other title',
-          content: 'real response',
-        });
-
-        const { result, waitForNextUpdate } = renderRestHook(
-          () => {
-            const create = useFetcher(ArticleResourceWithOtherListUrl.create());
-            const listA = useCache(ArticleResourceWithOtherListUrl.list(), {});
-            const listB = useCache(
-              ArticleResourceWithOtherListUrl.otherList(),
-              {},
-            );
-            return { create, listA, listB };
-          },
-          {
-            results: [
-              {
-                request: ArticleResourceWithOtherListUrl.otherList(),
-                params: {},
-                result: [{ id: 100, content: 'something' }],
-              },
-            ],
-          },
-        );
-
-        expect(result.current.listA).toEqual(undefined);
-        expect(result.current.listB).toEqual([existingItem]);
-
-        const promise = result.current.create({}, body, [
-          [
-            ArticleResourceWithOtherListUrl.list(),
-            {},
-            (newArticleID: string, articleIDs: string[] | undefined) => [
-              ...(articleIDs || []),
-              newArticleID,
-            ],
-          ],
-          [
-            ArticleResourceWithOtherListUrl.otherList(),
-            {},
-            (newArticleID: string, articleIDs: string[] | undefined) => [
-              ...(articleIDs || []),
-              newArticleID,
-            ],
-          ],
-        ]);
-
-        expect(result.current.listA).toEqual([
-          CoolerArticleResource.fromJS(body),
-        ]);
-        expect(result.current.listB).toEqual([
-          existingItem,
-          CoolerArticleResource.fromJS(body),
-        ]);
-        await promise;
-        expect(result.current.listA).toEqual([
-          CoolerArticleResource.fromJS({
-            ...payload,
-            title: 'some other title',
-            content: 'real response',
-          }),
-        ]);
-        expect(result.current.listB).toEqual([
-          existingItem,
-          CoolerArticleResource.fromJS({
-            ...payload,
-            title: 'some other title',
-            content: 'real response',
-          }),
-        ]);
-      });
-
-      it('should update on create', async () => {
-        const { result, waitForNextUpdate } = renderRestHook(() => {
-          const articles = useResource(FutureArticleResource.list(), {});
-          const create = useFetcher(FutureArticleResource.create());
-          return { articles, create };
-        });
-
-        await waitForNextUpdate();
-        expect(result.current.articles.map(({ id }) => id)).toEqual([5, 3]);
-
-        await result.current.create({
-          id: 1,
-        });
-        expect(result.current.articles.map(({ id }) => id)).toEqual([1, 5, 3]);
-      });
-
-      it('should clear only earlier optimistic updates when a promise resolves', async () => {
-        jest.useFakeTimers();
-
-        const params = { id: payload.id };
-        const { result, waitForNextUpdate } = renderRestHook(
-          () => {
-            const put = useFetcher(CoolerArticleResource.partialUpdate());
-            const article = useCache(CoolerArticleResource.detail(), params);
-            return { put, article };
-          },
-          {
-            results: [
-              {
-                request: CoolerArticleResource.detail(),
-                params,
-                result: payload,
-              },
-            ],
-          },
-        );
-
-        // first optimistic
-        mynock
-          .patch('/article-cooler/5')
-          .delay(200)
-          .reply(200, {
-            ...payload,
-            title: 'first',
-            content: 'first',
-          });
-        result.current.put(params, {
-          title: 'firstoptimistic',
-          content: 'firstoptimistic',
-        });
-        expect(result.current.article).toEqual(
-          CoolerArticleResource.fromJS({
-            ...payload,
-            title: 'firstoptimistic',
-            content: 'firstoptimistic',
-          }),
-        );
-
-        // second optimistic
-        mynock
-          .patch('/article-cooler/5')
-          .delay(50)
-          .reply(200, {
-            ...payload,
-            title: 'second',
-          });
-        result.current.put(params, {
-          title: 'secondoptimistic',
-        });
-        expect(result.current.article).toEqual(
-          CoolerArticleResource.fromJS({
-            ...payload,
-            title: 'secondoptimistic',
-            content: 'firstoptimistic',
-          }),
-        );
-
-        // third optimistic
-        mynock
-          .patch('/article-cooler/5')
-          .delay(500)
-          .reply(200, {
-            ...payload,
-            tags: ['third'],
-          });
-        result.current.put(params, {
-          tags: ['thirdoptimistic'],
-        });
-        expect(result.current.article).toEqual(
-          CoolerArticleResource.fromJS({
-            ...payload,
-            title: 'secondoptimistic',
-            content: 'firstoptimistic',
-            tags: ['thirdoptimistic'],
-          }),
-        );
-
-        // resolve second request while first is in flight
-        jest.advanceTimersByTime(51);
-        await waitForNextUpdate();
-
-        // first and second optimistic should be cleared with only third optimistic left to be layerd
-        // on top of second's network response
-        expect(result.current.article).toEqual(
-          CoolerArticleResource.fromJS({
-            ...payload,
-            title: 'second',
-          }),
-        );
-      });
-
-      describe('indexes', () => {
-        it('should resolve parallel useResource() request', async () => {
-          const { result, waitForNextUpdate } = renderRestHook(() => {
-            useResource(IndexedUserResource.list(), {});
-            return {
-              bob: useCache(IndexedUserResource.index(), {
-                username: 'bob',
-              }),
-              charlie: useCache(IndexedUserResource.index(), {
-                username: 'charlie',
-              }),
-              fetch: useFetcher(IndexedUserResource.detail()),
-              del: useFetcher(IndexedUserResource.delete()),
-              state: useContext(StateContext),
-            };
-          });
-          expect(result.current).toBeUndefined();
-          await waitForNextUpdate();
-          const bob = result.current.bob;
-          expect(bob).toBeDefined();
-          expect(bob instanceof UserResource).toBe(true);
-          expect(bob?.username).toBe('bob');
-          expect(bob).toMatchSnapshot();
-          expect(result.current.charlie).toBeUndefined();
-
-          const renamed = { ...users[0] };
-          renamed.username = 'charlie';
-          mynock.get(`/user/23`).reply(200, renamed);
-          result.current.fetch({ id: '23' });
-          await waitForNextUpdate();
-
-          const charlie = result.current.charlie;
-          expect(charlie).toBeDefined();
-          expect(charlie instanceof UserResource).toBe(true);
-          expect(charlie?.username).toBe('charlie');
-          expect(charlie).toMatchSnapshot();
-          expect(result.current.bob).toBeUndefined();
-
-          // deletes should roll through to indexes
-          await result.current.del({ id: '23' });
-          expect(result.current.charlie).toBeUndefined();
-        });
-      });
-    });
   });
 }
