+++ packages/core/src/react-integration/__tests__/subscriptions.tsx
@@ -90,6 +90,7 @@ for (const makeProvider of [makeCacheProvider, makeExternalCacheProvider]) {
         .reply(200, { ...articlePayload, title: 'sixer' });
       jest.advanceTimersByTime(frequency);
       expect((result.current as any).title).toBe('fiver');
+      jest.useRealTimers();
     });
 
     it('useSubscription() without active arg', async () => {
@@ -109,6 +110,7 @@ for (const makeProvider of [makeCacheProvider, makeExternalCacheProvider]) {
         waitForNextUpdate,
         articlePayload,
       );
+      jest.useRealTimers();
     });
 
     it('useSubscription() should dispatch rest-hooks/subscribe only once even with rerender', () => {
