+++ packages/core/src/react-integration/__tests__/useResource.web.tsx
@@ -224,6 +224,7 @@ describe('useResource()', () => {
     `);
     // eslint-disable-next-line require-atomic-updates
     console.error = oldError;
+    jest.useRealTimers();
   });
   it('should NOT suspend if result already in cache and options.invalidIfStale is false', () => {
     const state: State<unknown> = mockInitialState([
