+++ packages/core/src/state/__tests__/RIC.web.ts
@@ -11,5 +11,6 @@ describe('RequestIdleCallback', () => {
     jest.runAllTimers();
     expect(fn).toBeCalled();
     (global as any).requestIdleCallback = requestIdle;
+    jest.useRealTimers();
   });
 });
